refactor(optimizer): move rule/utils
